allow customization of onnx runtime version with env variable
the library depends on the version of onnx runtime installed on the
device. However, some devices might have a higher version installed
already. The version was hardcoded in the code making it hard for anyone
to choose the version they already have. With this fix, we allow the
customization of the onnx runtime version by setting it with an env
variable called CHROMAGO_ONNX_RUNTIME_VERSION. All dependent variables
are also derived from it.